Rework API by switching to blueprints
