+++ backend_py/src/blueprints/__init__.py
@@ -1,0 +1,4 @@
+from .cameras import *
+from .lights import *
+from .logs import *
+from .preferences import *
