+++ backend_py/src/logging/__init__.py
@@ -1,0 +1,1 @@
+from .log_handler import *
